# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type StreamStartInputAspectRatio = typeof StreamStartInputAspectRatio[keyof typeof StreamStartInputAspectRatio];


export const StreamStartInputAspectRatio = {
  shorts: 'shorts',
  full: 'full',
  square: 'square',
} as const;

export type StreamStartInputFacePosition = typeof StreamStartInputFacePosition[keyof typeof StreamStartInputFacePosition];


export const StreamStartInputFacePosition = {
  'top-left': 'top-left',
  'top-right': 'top-right',
  'bottom-left': 'bottom-left',
  'bottom-right': 'bottom-right',
  center: 'center',
} as const;

export interface StreamStartInput {
  /** @minLength 1 */
  streamId: string;
  /** @minLength 1 */
  ingestUrl: string;
  /** @minLength 1 */
  category: string;
  videoSource?: string;
  /** @items.minLength 1 */
  videoSources?: string[];
  faceCategory?: string;
  faceSource?: string;
  faceSources?: string[];
  /**
     * @minimum 0.5
     * @maximum 2
     */
  playbackSpeed?: number;
  aspectRatio?: StreamStartInputAspectRatio;
  facePosition?: StreamStartInputFacePosition;
  /**
     * @minimum 0.1
     * @maximum 0.6
     */
  faceScale?: number;
  /**
     * @minimum 1
     * @maximum 1440
     */
  durationMinutes?: number;
  autoRestart?: boolean;
}

export interface StreamStopInput {
  /** @minLength 1 */
  streamId: string;
}

export type StreamControlResponseStatus = typeof StreamControlResponseStatus[keyof typeof StreamControlResponseStatus];


export const StreamControlResponseStatus = {
  running: 'running',
  stopped: 'stopped',
  failed: 'failed',
} as const;

export interface StreamControlResponse {
  streamId: string;
  status: StreamControlResponseStatus;
  message: string;
  /** @nullable */
  pid: number | null;
}

export interface MediaUploadResponse {
  fileId: string;
  filename: string;
  sourcePath: string;
  playbackUrl: string;
}

export interface YoutubeDownloadInput {
  /** @minLength 1 */
  url: string;
}

export interface YoutubeDownloadResponse {
  fileId: string;
  filename: string;
  sourcePath: string;
  playbackUrl: string;
  title: string;
  duration: string;
}

export interface CreateLicenseInput {
  /** @minLength 1 */
  name: string;
  /**
     * @minimum 1
     * @maximum 3650
     */
  days?: number;
}

export interface RenewLicenseInput {
  /**
     * @minimum 1
     * @maximum 3650
     */
  days?: number;
}

export interface LicenseClientInput {
  /** @minLength 8 */
  key: string;
  /** @minLength 8 */
  clientId: string;
}

export type RenewLicenseForUserInput = LicenseClientInput & {
  /**
     * @minimum 1
     * @maximum 3650
     */
  days?: number;
};

export interface LicenseAccess {
  licenseId: string;
  key: string;
  name: string;
  expiresAt: string;
  active: boolean;
  clientId?: string;
}

export interface LicenseListResponse {
  licenses: LicenseAccess[];
}

export interface DeleteLicenseResult {
  licenseId: string;
  deleted: boolean;
}

export interface LicenseWorkspaceData { [key: string]: unknown }

export interface LicenseWorkspaceResponse {
  license: LicenseAccess;
  data: LicenseWorkspaceData | null;
}

export type SaveLicenseWorkspaceInput = LicenseClientInput & {
  data: LicenseWorkspaceData;
};

export interface WorkspaceSaveResponse {
  saved: boolean;
}

export type OwnerPasswordParameter = string;