# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: streaming
    description: Live stream control operations
  - name: media
    description: Local media upload operations
  - name: licenses
    description: License access and workspace operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /stream/start:
    post:
      operationId: startStream
      tags: [streaming]
      summary: Start a live stream
      description: Starts FFmpeg and sends a selected local category video to the supplied ingest URL.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StreamStartInput"
      responses:
        "202":
          description: Stream process accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreamControlResponse"
        "400":
          description: Invalid stream configuration
        "409":
          description: Stream is already running
        "500":
          description: Stream process could not start
  /stream/stop:
    post:
      operationId: stopStream
      tags: [streaming]
      summary: Stop a live stream
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StreamStopInput"
      responses:
        "200":
          description: Stream stopped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreamControlResponse"
        "404":
          description: Stream is not running
  /stream/status/{streamId}:
    get:
      operationId: getStreamStatus
      tags: [streaming]
      summary: Get stream process status
      parameters:
        - name: streamId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Stream status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreamControlResponse"
  /media/upload:
    post:
      operationId: uploadMedia
      tags: [media]
      summary: Upload a local video for server-side streaming
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
          video/*:
            schema:
              type: string
              format: binary
      parameters:
        - name: X-File-Name
          in: header
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        "201":
          description: Media is ready for streaming
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaUploadResponse"
        "400":
          description: Invalid media upload
  /media/files/{fileId}:
    get:
      operationId: getMediaFile
      tags: [media]
      summary: Serve an uploaded local media file
      parameters:
        - name: fileId
          in: path
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: Media file
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "404":
          description: Media file not found
  /media/youtube-download:
    post:
      operationId: downloadYoutubeVideo
      tags: [media]
      summary: Download a YouTube video at the best available quality
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/YoutubeDownloadInput"
      responses:
        "201":
          description: YouTube video downloaded and ready for streaming
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/YoutubeDownloadResponse"
        "400":
          description: Invalid URL or download failure
  /licenses:
    get:
      operationId: listLicenses
      tags: [licenses]
      summary: List licenses for the owner
      parameters:
        - $ref: "#/components/parameters/OwnerPassword"
      responses:
        "200":
          description: License list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LicenseListResponse"
        "401":
          description: Owner password is incorrect
    post:
      operationId: createLicense
      tags: [licenses]
      summary: Create a license
      parameters:
        - $ref: "#/components/parameters/OwnerPassword"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLicenseInput"
      responses:
        "201":
          description: License created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LicenseAccess"
        "401":
          description: Owner password is incorrect
  /licenses/{licenseId}:
    delete:
      operationId: deleteLicense
      tags: [licenses]
      summary: Delete a license and its workspace data
      parameters:
        - $ref: "#/components/parameters/OwnerPassword"
        - name: licenseId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: License deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteLicenseResult"
        "401":
          description: Owner password is incorrect
  /licenses/{licenseId}/renew:
    post:
      operationId: renewLicense
      tags: [licenses]
      summary: Renew a license as the owner
      parameters:
        - $ref: "#/components/parameters/OwnerPassword"
        - name: licenseId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RenewLicenseInput"
      responses:
        "200":
          description: License renewed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LicenseAccess"
        "401":
          description: Owner password is incorrect
  /licenses/validate:
    post:
      operationId: validateLicense
      tags: [licenses]
      summary: Validate a license key for a browser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LicenseClientInput"
      responses:
        "200":
          description: License is active
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LicenseAccess"
        "403":
          description: License is invalid or expired
  /licenses/workspace/get:
    post:
      operationId: getLicenseWorkspace
      tags: [licenses]
      summary: Load a browser-specific license workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LicenseClientInput"
      responses:
        "200":
          description: Workspace loaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LicenseWorkspaceResponse"
        "403":
          description: License is invalid or expired
  /licenses/workspace:
    put:
      operationId: saveLicenseWorkspace
      tags: [licenses]
      summary: Save a browser-specific license workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveLicenseWorkspaceInput"
      responses:
        "200":
          description: Workspace saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceSaveResponse"
        "403":
          description: License is invalid or expired
  /licenses/renew:
    post:
      operationId: renewLicenseForUser
      tags: [licenses]
      summary: Renew a license from the user side
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RenewLicenseForUserInput"
      responses:
        "200":
          description: License renewed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LicenseAccess"
        "404":
          description: License key not found
components:
  parameters:
    OwnerPassword:
      name: X-Owner-Password
      in: header
      required: true
      schema:
        type: string
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    StreamStartInput:
      type: object
      properties:
        streamId:
          type: string
          minLength: 1
        ingestUrl:
          type: string
          minLength: 1
        category:
          type: string
          minLength: 1
        videoSource:
          type: string
        videoSources:
          type: array
          items:
            type: string
            minLength: 1
        faceCategory:
          type: string
        faceSource:
          type: string
        faceSources:
          type: array
          items:
            type: string
        playbackSpeed:
          type: number
          minimum: 0.5
          maximum: 2
          default: 1
        aspectRatio:
          type: string
          enum: [shorts, full, square]
          default: full
        facePosition:
          type: string
          enum: [top-left, top-right, bottom-left, bottom-right, center]
          default: bottom-right
        faceScale:
          type: number
          minimum: 0.1
          maximum: 0.6
          default: 0.25
        durationMinutes:
          type: number
          minimum: 1
          maximum: 1440
        autoRestart:
          type: boolean
          default: false
      required:
        - streamId
        - ingestUrl
        - category
    StreamStopInput:
      type: object
      properties:
        streamId:
          type: string
          minLength: 1
      required:
        - streamId
    StreamControlResponse:
      type: object
      properties:
        streamId:
          type: string
        status:
          type: string
          enum: [running, stopped, failed]
        message:
          type: string
        pid:
          type: ["number", "null"]
      required:
        - streamId
        - status
        - message
        - pid
    MediaUploadResponse:
      type: object
      properties:
        fileId:
          type: string
        filename:
          type: string
        sourcePath:
          type: string
        playbackUrl:
          type: string
      required:
        - fileId
        - filename
        - sourcePath
        - playbackUrl
    YoutubeDownloadInput:
      type: object
      properties:
        url:
          type: string
          minLength: 1
      required:
        - url
    YoutubeDownloadResponse:
      type: object
      properties:
        fileId:
          type: string
        filename:
          type: string
        sourcePath:
          type: string
        playbackUrl:
          type: string
        title:
          type: string
        duration:
          type: string
      required:
        - fileId
        - filename
        - sourcePath
        - playbackUrl
        - title
        - duration
    CreateLicenseInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        days:
          type: number
          minimum: 1
          maximum: 3650
          default: 30
      required:
        - name
    RenewLicenseInput:
      type: object
      properties:
        days:
          type: number
          minimum: 1
          maximum: 3650
          default: 30
    LicenseClientInput:
      type: object
      properties:
        key:
          type: string
          minLength: 8
        clientId:
          type: string
          minLength: 8
      required:
        - key
        - clientId
    RenewLicenseForUserInput:
      allOf:
        - $ref: "#/components/schemas/LicenseClientInput"
        - type: object
          properties:
            days:
              type: number
              minimum: 1
              maximum: 3650
              default: 30
    LicenseAccess:
      type: object
      properties:
        licenseId:
          type: string
        key:
          type: string
        name:
          type: string
        expiresAt:
          type: string
          format: date-time
        active:
          type: boolean
        clientId:
          type: string
      required:
        - licenseId
        - key
        - name
        - expiresAt
        - active
    LicenseListResponse:
      type: object
      properties:
        licenses:
          type: array
          items:
            $ref: "#/components/schemas/LicenseAccess"
      required:
        - licenses
    DeleteLicenseResult:
      type: object
      properties:
        licenseId:
          type: string
        deleted:
          type: boolean
      required:
        - licenseId
        - deleted
    LicenseWorkspaceData:
      type: object
      additionalProperties: true
    LicenseWorkspaceResponse:
      type: object
      properties:
        license:
          $ref: "#/components/schemas/LicenseAccess"
        data:
          oneOf:
            - $ref: "#/components/schemas/LicenseWorkspaceData"
            - type: "null"
      required:
        - license
        - data
    SaveLicenseWorkspaceInput:
      allOf:
        - $ref: "#/components/schemas/LicenseClientInput"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/LicenseWorkspaceData"
          required:
            - data
    WorkspaceSaveResponse:
      type: object
      properties:
        saved:
          type: boolean
      required:
        - saved