# lib/api-zod/src/generated/types/createLicenseInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateLicenseInput {
  /** @minLength 1 */
  name: string;
  /**
     * @minimum 1
     * @maximum 3650
     */
  days?: number;
}