# lib/api-zod/src/generated/types/licenseAccess.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LicenseAccess {
  licenseId: string;
  key: string;
  name: string;
  expiresAt: Date;
  active: boolean;
  clientId?: string;
}