# lib/api-zod/src/generated/types/licenseClientInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LicenseClientInput {
  /** @minLength 8 */
  key: string;
  /** @minLength 8 */
  clientId: string;
}