# lib/api-zod/src/generated/types/licenseListResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LicenseAccess } from './licenseAccess';

export interface LicenseListResponse {
  licenses: LicenseAccess[];
}