# lib/api-zod/src/generated/types/licenseWorkspaceResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LicenseAccess } from './licenseAccess';
import type { LicenseWorkspaceData } from './licenseWorkspaceData';

export interface LicenseWorkspaceResponse {
  license: LicenseAccess;
  data: LicenseWorkspaceData | null;
}