# lib/api-zod/src/generated/types/renewLicenseForUserInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LicenseClientInput } from './licenseClientInput';

export type RenewLicenseForUserInput = LicenseClientInput & {
  /**
     * @minimum 1
     * @maximum 3650
     */
  days?: number;
};