# lib/api-zod/src/generated/types/renewLicenseInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RenewLicenseInput {
  /**
     * @minimum 1
     * @maximum 3650
     */
  days?: number;
}