# lib/api-zod/src/generated/types/saveLicenseWorkspaceInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LicenseClientInput } from './licenseClientInput';
import type { LicenseWorkspaceData } from './licenseWorkspaceData';

export type SaveLicenseWorkspaceInput = LicenseClientInput & {
  data: LicenseWorkspaceData;
};